Fix copy/paste error
...that had caused unoidl-write to generate garbage values when writing
constants of type double to binary type libraries.  (But which had apparently
gone largely unnoticed until now, as the only actual use so far of such a
constant of type double that I could find is

>     const double            aConstDouble    = 3.14e-10;

in the unused stoc/test/testcorefl.idl test file.)

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/164096
Tested-by: Jenkins

+++ unoidl/source/unoidl-write.cxx
@@ -188,7 +188,7 @@ void writeIso60599Binary32(osl::File & file, float value) {
 void writeIso60599Binary64(osl::File & file, double value) {
     union {
         unsigned char buf[8];
-        float d; // assuming double is ISO 60599 binary64
+        double d; // assuming double is ISO 60599 binary64
     } sa;
     sa.d = value;
 #if defined OSL_BIGENDIAN
